@@ -4,7 +4,6 @@
     "description": "Ramsey uuid generator implementation",
     "license": "MIT",
     "keywords": [
-        "qlimix",
         "uuid",
         "ramsey"
     ],
@@ -15,9 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "qlimix/uuid": "^1.0",
-        "ramsey/uuid": "^3.8"
+        "php": ">=7.4",
+        "qlimix/uuid": "v2.x-dev",
+        "ramsey/uuid": "^4.0"
+    },
+    "require-dev": {
+        "qlimix/code-standard": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -31,8 +33,5 @@
     },
     "config": {
         "sort-packages": true
-    },
-    "require-dev": {
-        "qlimix/code-standard": "^1.0"
     }
 }
